requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
distro==1.4.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
future==0.18.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
mpmath==1.1.0
numpy==1.17.3
pandas==0.25.3
psycopg2==2.8.4
pycurl==7.43.0.3
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.15
sympy==1.4
tabula==1.0.5
tabula-py==1.4.1
urllib3==1.25.6
Werkzeug==0.16.0
Wrapper==1.1.0b1
